new changes

+++ index.js
@@ -1,23 +1,17 @@
 console.log("start...");
 /* Navigation between pages */
 
-setTimeout(() => {
-    // Navigate to the new page after a delay of 1 second
+const signupBtn = document.querySelector(".signup");
+const loginBtn = document.querySelector(".login");
+
+// go to signUp page
+signupBtn.addEventListener("click", () => {
     window.location.assign("./index.html");
-    console.log("end...");
-}, 1000);
-
-// const signupBtn = document.querySelector(".signup");
-// const loginBtn = document.querySelector(".login");
-
-// // go to signUp page
-// signupBtn.addEventListener("click", () => {
-//     window.location.assign("./index.html");
-// });
-// //go to login page
-// loginBtn.addEventListener("click", () => {
-//     window.location.assign("./login.html");
-// });
+});
+//go to login page
+loginBtn.addEventListener("click", () => {
+    window.location.assign("./login.html");
+});
 
 // all Existing Users
 let existingUsers = [];
@@ -41,19 +35,23 @@ signUpForm.addEventListener("submit", (event) => {
     const pass = document.querySelector("#pass").value;
     const confirmPass = document.querySelector("#confirm-pass").value;
 
-    // getting previous users
-    let prevUsers = JSON.parse(localStorage.getItem("allUserInfo")) || [];
+    if (pass === confirmPass) {
+        // getting previous users
+        let prevUsers = JSON.parse(localStorage.getItem("allUserInfo")) || [];
 
-    // combining current + previous users
-    prevUsers.push({ username, email, pass, confirmPass });
+        // combining current + previous users
+        prevUsers.push({ username, email, pass, confirmPass });
 
-    // storing user credentials in local storage
-    localStorage.setItem("allUserInfo", JSON.stringify(prevUsers));
+        // storing user credentials in local storage
+        localStorage.setItem("allUserInfo", JSON.stringify(prevUsers));
 
-    console.log("user added...");
-
-    // clearing form
-    signUpForm.reset();
+        console.log("user added...");
+        // clearing form
+        signUpForm.reset();
+    } else {
+        console.log("pass and confirmPass must be same...");
+    }
 });
 
 console.log("end...");
+//
